VB-3437 - move endpoint to VisitNotificationController.
